copying file visitor moved
